EegProcessingUtil: Adjust samples verification
The check made earlier didn't really check if the difference between
two samples was a result of the brain being active or calm.

Modify differences from (max - min) to (last sample - first sample)
Then add an additional check to see if the difference is positive or
negative. Active is defined as positive and calm as negative.